requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy<2.0.0
torch==2.2.0
transformers @ git+https://github.com/QwenLM/transformers.git
qwen-vl-utils==0.0.10
python-dotenv==1.0.0
accelerate 